+++ jmccc-yggdrasil-authenticator/src/main/java/org/to2mbn/jmccc/auth/yggdrasil/core/util/Base64.java
@@ -1,4 +1,4 @@
-package org.to2mbn.jyal.util;
+package org.to2mbn.jmccc.auth.yggdrasil.core.util;
 
 public final class Base64 {
