+++ jmccc-yggdrasil-authenticator/src/main/java/org/to2mbn/jmccc/auth/yggdrasil/core/util/UUIDUtils.java
@@ -1,4 +1,4 @@
-package org.to2mbn.jyal.util;
+package org.to2mbn.jmccc.auth.yggdrasil.core.util;
 
 import java.util.UUID;
